+++ R/lazyframe__lazy.R
@@ -315,11 +315,6 @@ LazyFrame_with_row_index = function(name, offset = NULL) {
     unwrap("in $with_row_index():")
 }
 
-LazyFrame_with_row_count = function(name, offset = NULL) {
-  warning("`$with_row_count()` is deprecated and will be removed in 0.15.0. Use `with_row_index()` instead.")
-  .pr$LazyFrame$with_row_index(self, name, offset) |>
-    unwrap("in $with_row_count():")
-}
 
 #' Apply filter to LazyFrame
 #'
@@ -1908,7 +1903,9 @@ LazyFrame_rolling = function(
 #' )$collect()
 #'
 #' # Dynamic group bys can also be combined with grouping on normal keys
-#' lf = lf$with_columns(groups = pl$Series(c("a", "a", "a", "b", "b", "a", "a")))
+#' lf = lf$with_columns(
+#'   groups = as_polars_series(c("a", "a", "a", "b", "b", "a", "a"))
+#' )
 #' lf$collect()
 #'
 #' lf$group_by_dynamic(